draft-api: Move grep migration endpoint to public controller
