+++ draft-api/src/main/scala/no/ndla/draftapi/controller/DraftController.scala
@@ -26,7 +26,7 @@ import no.ndla.mapping
 import no.ndla.mapping.LicenseDefinition
 import no.ndla.network.tapir.NoNullJsonPrinter.*
 import no.ndla.network.tapir.TapirUtil.errorOutputsFor
-import no.ndla.network.tapir.auth.Permission.DRAFT_API_WRITE
+import no.ndla.network.tapir.auth.Permission.{ARTICLE_API_WRITE, DRAFT_API_WRITE}
 import no.ndla.network.tapir.{DynamicHeaders, TapirController}
 import sttp.model.StatusCode
 import sttp.tapir.*
@@ -122,7 +122,8 @@ trait DraftController {
       partialPublish,
       partialPublishMultiple,
       copyRevisionDates,
-      getArticleBySlug
+      getArticleBySlug,
+      migrateOutdatedGreps
     )
 
     /** Does a scroll with [[ArticleSearchService]] If no scrollId is specified execute the function @orFunction in the
@@ -677,6 +678,17 @@ trait DraftController {
           if (permitted) article
           else ErrorHelpers.forbidden.asLeft
         }
+      }
+
+    def migrateOutdatedGreps: ServerEndpoint[Any, Eff] = endpoint.post
+      .in("migrate-greps")
+      .summary("Iterate all articles and migrate outdated grep codes")
+      .description("Iterate all articles and migrate outdated grep codes")
+      .errorOut(errorOutputsFor(500))
+      .out(emptyOutput)
+      .requirePermission(DRAFT_API_WRITE, ARTICLE_API_WRITE)
+      .serverLogicPure { user => _ =>
+        writeService.migrateOutdatedGreps(user).handleErrorsOrOk
       }
   }
 }
