+++ draft-api/src/main/scala/no/ndla/draftapi/controller/InternController.scala
@@ -23,7 +23,7 @@ import com.typesafe.scalalogging.StrictLogging
 import no.ndla.network.tapir.NoNullJsonPrinter.*
 import no.ndla.network.tapir.TapirUtil.errorOutputsFor
 import no.ndla.network.tapir.TapirController
-import no.ndla.network.tapir.auth.Permission.{ARTICLE_API_WRITE, DRAFT_API_WRITE}
+import no.ndla.network.tapir.auth.Permission.DRAFT_API_WRITE
 import no.ndla.network.tapir.auth.TokenUser
 import scalikejdbc.ReadOnlyAutoSession
 import sttp.model.StatusCode
@@ -91,8 +91,7 @@ trait InternController {
       deleteArticle,
       dumpArticles,
       dumpSingleArticle,
-      postDump,
-      migrateOutdatedGreps
+      postDump
     )
 
     def postIndex: ServerEndpoint[Any, Eff] = endpoint.post
@@ -261,15 +260,6 @@ trait InternController {
       .out(jsonBody[Draft])
       .serverLogicPure { article =>
         writeService.insertDump(article)
-      }
-
-    def migrateOutdatedGreps: ServerEndpoint[Any, Eff] = endpoint.post
-      .in("migrate-greps")
-      .errorOut(errorOutputsFor(500))
-      .out(emptyOutput)
-      .requirePermission(ARTICLE_API_WRITE)
-      .serverLogicPure { user => _ =>
-        writeService.migrateOutdatedGreps(user).handleErrorsOrOk
       }
   }
 }
